+++ user_guide_src/source/changelog.rst
@@ -57,6 +57,8 @@ Release Date: Not Released
    -  Added Interbase/Firebird database support via the "interbase" driver
    -  Added an optional database name parameter to db_select().
    -  Replaced the _error_message() and _error_number() methods with error(), that returns an array containing the last database error code and message.
+   -  Improved version() implementation so that drivers that have a native function to get the version number don't have to be defined in the core DB_driver class.
+   -  PostgreSQL driver now uses pg_version() to get the database version number, when possible.
 
 -  Libraries